More updates than I expected...
I was hoping to just update the Simple Linear Regression page, but this seems to have updated other things. I recently updated my R and RStudio, so some of these updates may be because of that... Not sure what else changed. I guess I forgot to push for a while...